Phase_III/backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "Backend API for Phase 2 Todo Application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.0",
    "sqlmodel>=0.0.14",
    "psycopg[binary]>=3.1.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.0",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.46",
    "aiosqlite>=0.22.1",
    "asyncpg>=0.31.0",
    "httpx>=0.27.0",
    "pyjwt>=2.8.0",
    "cryptography>=41.0.0",
    "mcp>=1.0.0",
    "openai[agents]>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "aiosqlite>=0.19.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
